Moving leftovers into EKS
After recent refactoring all EKS related logic
is extracted into separate class. This patch
moving what was missed during refactoring into
right place
