+++ ocw/lib/eks.py
@@ -1,7 +1,10 @@
 import os
 import json
+import time
+from datetime import datetime, timedelta, timezone
 import kubernetes
 import boto3
+from webui.PCWConfig import PCWConfig, ConfigFile
 from ocw.lib.provider import Provider
 from ocw.lib.k8s import clean_jobs
 
@@ -11,13 +14,13 @@ TAG_IGNORE = 'pcw_ignore'
 class EKS(Provider):
     __instances = {}
     default_region: str = 'eu-central-1'
+    __cluster_regions = []
 
     def __new__(cls, vault_namespace):
         if vault_namespace not in EKS.__instances:
             EKS.__instances[vault_namespace] = self = object.__new__(cls)
             self.__eks_client = {}
             self.__kubectl_client = {}
-            self.__cluster_regions = None
             self.__aws_dir = None
 
         return EKS.__instances[vault_namespace]
@@ -25,18 +28,17 @@ class EKS(Provider):
     def __init__(self, namespace: str):
         super().__init__(namespace)
         self.create_credentials_file()
+        if len(EKS.__cluster_regions) == 0:
+            if PCWConfig.has('clusters/ec2_regions'):
+                EKS.__cluster_regions = ConfigFile().getList('clusters/ec2_regions')
+            else:
+                regions_query = self.cmd_exec(f"aws ec2 describe-regions --query 'Regions[].RegionName' --output json --region {EKS.default_region}")
+                EKS.__cluster_regions = json.loads(regions_query.stdout)
 
     def aws_dir(self):
         if self.__aws_dir is None:
             self.__aws_dir = os.path.expanduser("~/.aws")
         return self.__aws_dir
-
-    def list_regions(self):
-        if self.__cluster_regions is None:
-            regions_query = self.cmd_exec(
-                f"aws ec2 describe-regions --query 'Regions[].RegionName' --output json --region {EKS.default_region}")
-            self.__cluster_regions = json.loads(regions_query.stdout)
-        return self.__cluster_regions
 
     def create_credentials_file(self) -> None:
         creds_file = f"{self.aws_dir()}/credentials"
@@ -82,7 +84,7 @@ class EKS(Provider):
 
     def all_clusters(self) -> dict:
         clusters = {}
-        for region in self.list_regions():
+        for region in EKS.__cluster_regions:
             self.log_dbg("Checking clusters in {}", region)
             response = self.eks_client(region).list_clusters()
             if 'clusters' in response and len(response['clusters']) > 0:
@@ -98,8 +100,47 @@ class EKS(Provider):
                     del clusters[region]
         return clusters
 
+    def wait_for_empty_nodegroup_list(self, region: str, cluster_name: str, timeout_minutes: int = 20):
+        if self.dry_run:
+            self.log_info("Skip waiting due to dry-run mode")
+            return None
+        self.log_dbg("Waiting empty nodegroup list in {}", cluster_name)
+        end = datetime.now(timezone.utc) + timedelta(minutes=timeout_minutes)
+        resp_nodegroup = self.eks_client(region).list_nodegroups(clusterName=cluster_name)
+
+        while datetime.now(timezone.utc) < end and len(resp_nodegroup['nodegroups']) > 0:
+            time.sleep(20)
+            resp_nodegroup = self.eks_client(region).list_nodegroups(clusterName=cluster_name)
+            if len(resp_nodegroup['nodegroups']) > 0:
+                self.log_dbg("Still waiting for {} nodegroups to disappear", len(resp_nodegroup['nodegroups']))
+        return None
+
+    def delete_all_clusters(self) -> None:
+        self.log_info("Deleting all clusters!")
+        for region in self.__cluster_regions:
+            response = self.eks_client(region).list_clusters()
+            if len(response['clusters']):
+                self.log_dbg("Found {} cluster(s) in {}", len(response['clusters']), region)
+                for cluster in response['clusters']:
+                    resp_nodegroup = self.eks_client(region).list_nodegroups(clusterName=cluster)
+                    if len(resp_nodegroup['nodegroups']):
+                        self.log_dbg("Found {} nodegroups for {}", len(resp_nodegroup['nodegroups']), cluster)
+                        for nodegroup in resp_nodegroup['nodegroups']:
+                            if self.dry_run:
+                                self.log_info("Skipping {} nodegroup deletion due to dry-run mode", nodegroup)
+                            else:
+                                self.log_info("Deleting {}", nodegroup)
+                                self.eks_client(region).delete_nodegroup(
+                                    clusterName=cluster, nodegroupName=nodegroup)
+                        self.wait_for_empty_nodegroup_list(region, cluster)
+                    if self.dry_run:
+                        self.log_info("Skipping {} cluster deletion due to dry-run mode", cluster)
+                    else:
+                        self.log_info("Finally deleting {} cluster", cluster)
+                        self.eks_client(region).delete_cluster(name=cluster)
+
     def cleanup_k8s_jobs(self):
-        for region in self.list_regions():
+        for region in self.__cluster_regions:
             self.log_dbg(f"Region {region}")
             clusters = self.eks_client(region).list_clusters()['clusters']
             for cluster_name in clusters:
